fixed a bunch of annoying errors, and attempted to fix the norm loss when calculating eigenvalues.

+++ common/common.hpp
@@ -612,8 +612,8 @@ namespace common
    Mat populate_matrix(const Parameters &params,
                        Test test,
                        FindValue find_value,
-                       const unsigned int mat_size_m,
-                       const unsigned int mat_size_n,
+                       const int mat_size_m,
+                       const int mat_size_n,
                        //const unsigned int diagonal_storage,
                        //const unsigned int offdiag_storage,
                        const bool symmetric=true,
@@ -644,11 +644,11 @@ namespace common
        PetscInt dnnz[rowend-rowstart];
        PetscInt onnz[rowend-rowstart];
        //find the preallocation functions:
-       for (size_t i = rowstart; i < rowend; i++)
+       for (int i = rowstart; i < rowend; i++)
        {
            dnnz[i-rowstart] = 0;
            onnz[i-rowstart] = 0;
-           for (size_t j = 0; j < mat_size_n; j++)
+           for (int j = 0; j < mat_size_n; j++)
            {
                if (test(i,j))
                {
